Fixed a small but with std::max.

+++ oss_src/unity/extensions/cumulative_aggregates.cpp
@@ -23,7 +23,7 @@ void check_vector_equal_size(const gl_sarray& in) {
   // Initialize. 
   DASSERT_TRUE(in.dtype() == flex_type_enum::VECTOR); 
   size_t n_threads = thread::cpu_count();
-  n_threads = std::max(n_threads, 1);
+  n_threads = std::max(n_threads, size_t(1));
   size_t m_size = in.size();
           
   // Throw the following error.
